Loosen package version restrictions and make into peer dependencies
In order to keep the package as small and flexible as possible I've
followed the current approach of externalizing all the dependencies.

In this commit I've added the 'create-react-class' and 'prop-types'
references to the peer dependencies of the package.

I've also loosened the version number restrictions to make this as widely
compatible as possible.